chore: move scripts from package.json to project.json targets
