+++ packages/certificate-summary/package.json
@@ -21,11 +21,5 @@
   },
   "bugs": {
     "url": "https://github.com/s1seven/schema-tools/issues"
-  },
-  "scripts": {
-    "clean": "rimraf -rf ./dist",
-    "cleanup": "rimraf -rf dist coverage",
-    "scan:deps": "audit-ci --config ../../audit-ci.json",
-    "example": "npm run build && node example.js"
   }
 }
