+++ packages/generate-coa-pdf-template/package.json
@@ -32,11 +32,5 @@
         "@s1seven/schema-tools-types": true
       }
     }
-  },
-  "scripts": {
-    "scan:deps": "audit-ci --config ../../audit-ci.json",
-    "create-example": "node ./utils/create-example.js",
-    "create-interfaces": "node ./utils/create-schema-interfaces.js https://schemas.s1seven.com/coa-schemas/v0.2.0/schema.json ./src/types/schemaTypes.ts",
-    "create-interfaces-interactive": "node ../../utils/create-schema-interfaces-interactive.js ./src/types/schemaTypes.ts"
   }
 }
