+++ packages/generate-en10168-pdf-template/project.json
@@ -34,6 +34,37 @@
         "lintFilePatterns": ["packages/generate-en10168-pdf-template/**/*.ts"]
       }
     },
+    "scan:deps": {
+      "executor": "nx:run-commands",
+      "options": {
+        "cwd": "packages/generate-en10168-pdf-template",
+        "command": "audit-ci --config ../../audit-ci.json"
+      }
+    },
+    "create-example": {
+      "executor": "nx:run-commands",
+      "options": {
+        "cwd": "packages/generate-en10168-pdf-template",
+        "command": "node --icu-data-dir=`echo $(./node_modules/.bin/node-full-icu-path)` ./utils/create-example.js"
+      },
+      "dependsOn": ["build"]
+    },
+    "create-example:no-icu": {
+      "executor": "nx:run-commands",
+      "options": {
+        "cwd": "packages/generate-en10168-pdf-template",
+        "command": "node ./utils/create-example.js"
+      },
+      "dependsOn": ["build"]
+    },
+    "create-interfaces": {
+      "executor": "nx:run-commands",
+      "options": {
+        "cwd": "packages/generate-en10168-pdf-template",
+        "command": "node ./utils/create-schema-interfaces.js https://schemas.s1seven.com/en10168-schemas/v0.2.0/schema.json ./src/types/schemaTypes.ts"
+      },
+      "dependsOn": ["build"]
+    },
     "test": {
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
