+++ packages/generate-html/example.js
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 
 const { writeFileSync } = require('fs');
-const { generateHtml } = require('./dist/index');
+const { generateHtml } = require('../../dist/generate-html/src/index');
 
 const defaultExternalCertificatePath = `${__dirname}/../../fixtures/EN10168/v0.2.0/valid_cert.json`;
 const defaultOutputPath = `${__dirname}/template.html`;
