+++ packages/generate-pdf-template-helpers/package.json
@@ -24,10 +24,5 @@
         "node": ">=12.x"
       }
     }
-  },
-  "scripts": {
-    "clean": "rimraf -rf ./dist",
-    "cleanup": "rimraf -rf dist coverage",
-    "scan:deps": "audit-ci --config ../../audit-ci.json"
   }
 }
