+++ packages/utils/package.json
@@ -24,11 +24,5 @@
   "engines": {
     "node": ">=12",
     "npm": ">=6"
-  },
-  "scripts": {
-    "clean": "rimraf -rf ./dist",
-    "cleanup": "rimraf -rf dist coverage",
-    "scan:deps": "audit-ci --config ../../audit-ci.json",
-    "cast-cert-test": "node example.js"
   }
 }
